Properly added VAELoss
