+++ clinicadl/utils/network/autoencoder/models.py
@@ -3,7 +3,7 @@ from torch import nn
 from clinicadl.utils.network.autoencoder.cnn_transformer import CNN_Transformer
 from clinicadl.utils.network.cnn.models import Conv4_FC3, Conv5_FC3, resnet18
 from clinicadl.utils.network.sub_network import AutoEncoder
-from clinicadl.utils.network.vae.vae_utils import (
+from clinicadl.utils.network.vae.vae_layers import (
     DecoderLayer3D,
     EncoderLayer3D,
     Flatten,
